requirements.txt
beautifulsoup4
cloudscraper
faiss-cpu
openai
sentence-transformers
selenium